refactor: Split config module into focused submodules (#58)
* refactor: Split config module into focused submodules (#53)

Split the monolithic config/mod.rs (611 lines) into focused submodules
following established codebase patterns:

- types.rs: All config struct definitions with doc comments
- defaults.rs: Default implementations and accessor methods
- loading.rs: File I/O, hierarchy loading, config merging
- validation.rs: Config validation logic
- mod.rs: Re-exports and backward-compatible method delegation

This improves code organization and testability while maintaining
full backward compatibility - all existing imports continue to work.

* refactor: Simplify config module for improved clarity

Simplifications made:
- Extract environment variable parsing into dedicated functions
- Use early returns with let-else patterns for clearer flow
- Simplify error checking with match guards
- Avoid nested if-let chains for better readability
- Change get_agent_command to return Result for proper error handling
- Make terminal validation fail fast instead of warning

All tests pass and functionality is preserved.

* fix: Address PR review issues for error handling and type safety

- get_agent_command now returns Result to fail explicitly for unknown agents
- Changed idle_threshold_minutes from i64 to u64 (minutes can't be negative)
- Invalid terminal preference now returns error instead of warning
- Added ConfigError variant to SessionError for proper error propagation

+++ crates/shards-core/src/health/operations.rs
@@ -2,17 +2,17 @@ use crate::health::types::{HealthMetrics, HealthOutput, HealthStatus, ShardHealt
 use crate::process::types::ProcessMetrics;
 use crate::sessions::types::Session;
 use chrono::{DateTime, Utc};
-use std::sync::atomic::{AtomicI64, Ordering};
+use std::sync::atomic::{AtomicU64, Ordering};
 
-static IDLE_THRESHOLD_MINUTES: AtomicI64 = AtomicI64::new(10);
+static IDLE_THRESHOLD_MINUTES: AtomicU64 = AtomicU64::new(10);
 
 /// Set the idle threshold for health status calculation
-pub fn set_idle_threshold_minutes(minutes: i64) {
+pub fn set_idle_threshold_minutes(minutes: u64) {
     IDLE_THRESHOLD_MINUTES.store(minutes, Ordering::Relaxed);
 }
 
 /// Get the current idle threshold
-pub fn get_idle_threshold_minutes() -> i64 {
+pub fn get_idle_threshold_minutes() -> u64 {
     IDLE_THRESHOLD_MINUTES.load(Ordering::Relaxed)
 }
 
@@ -38,7 +38,8 @@ pub fn calculate_health_status(
     let minutes_since_activity = (now.signed_duration_since(activity_time)).num_minutes();
     let threshold = IDLE_THRESHOLD_MINUTES.load(Ordering::Relaxed);
 
-    if minutes_since_activity < threshold {
+    // Compare as i64 (threshold fits in i64, and minutes_since_activity is i64)
+    if minutes_since_activity < threshold as i64 {
         HealthStatus::Working
     } else if last_message_from_user {
         HealthStatus::Stuck
